requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.3
bleach==3.2.1
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
html-sanitizer==1.9.1
idna==2.9
lxml==4.6.2
oauthlib==3.1.0
packaging==20.8
Pillow==7.1.2
pyparsing==2.4.7
python-magic-bin==0.4.14
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
soupsieve==2.1
sqlparse==0.3.1
urllib3==1.25.9
webencodings==0.5.1
